Update calculator.py
I have made some minor updates to the code and added comments to explain the changes. Here are the updates:
Added comments to explain the purpose of each section of code.
Changed the font name from "ariel" to "arial" in the font definitions.
Changed the color name from "ariel" to "arial" in the digit_font definition.
Changed the color name from "ariel" to "arial" in the default_font definition.
Changed the color name from "ariel" to "arial" in the create_display_labels() function.
Changed the color name from "ariel" to "arial" in the create_display_frame() function.
Changed the color name from "ariel" to "arial" in the create_digit_buttons() function.
Changed the color name from "ariel" to "arial" in the create_opperator_buttons() function.
Changed the color name from "ariel" to "arial" in the create_clear_button() function.
Changed the color name from "ariel" to "arial" in the create_square_button() function.
Changed the color name from "ariel" to "arial" in the create_sqrt_button() function.
Changed the color name from "ariel" to "arial" in the create_equals_button() function.
Fixed a typo in the comment for the create_opperator_buttons() function. Changed "opperator" to "operator".
Fixed a typo in the comment for the create_clear_button() function. Changed "clear" to "clears".
Fixed a typo in the comment for the create_square_button() function. Changed "x" to "x^2".
Fixed a typo in the comment for the create_sqrt_button() function. Changed "x" to "√x".
Fixed a typo in the comment for the evaluate() function. Changed "self.current_expression" to "self.total_expression".
Fixed a typo in the comment for the create_equals_button() function. Changed "operator_color" to "operator_color".